move security plugins to a sub ingest directory
It makes sense to introduce new Security ingest
processors (example: #31087), and this change would
give them a good place to be written.
