+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/security/ingest/SetSecurityUserProcessorFactoryTests.java
@@ -3,11 +3,11 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-package org.elasticsearch.xpack.core.security.authz.accesscontrol;
+package org.elasticsearch.xpack.core.security.ingest;
 
 import org.elasticsearch.ElasticsearchParseException;
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.core.security.authz.accesscontrol.SetSecurityUserProcessor.Property;
+import org.elasticsearch.xpack.core.security.ingest.SetSecurityUserProcessor.Property;
 
 import java.util.Arrays;
 import java.util.EnumSet;
